--- docs/goals.md
@@ -1,10 +1,0 @@
-goals:
-- quereable
-- equation system
-- symbolic execution
-- compile IR back to native code
-- transformations (instrumentation)
-- decompiler backend
-
-non-goals:
-... think some
