+++ retoolkit.cabal
@@ -1,24 +1,36 @@
 
- name:                retoolkit
- version:             0.1.0.0
- description:         another reverse engineering toolkit
- license:             GPL
- license-file:        LICENSE
- author:              SD
- maintainer:          ktion23@gmail.com
- build-type:          Simple
- cabal-version:       >=1.10
+name:                retoolkit
+version:             0.1.0.0
+description:         another reverse engineering toolkit
+license:             GPL
+license-file:        LICENSE
+author:              SD
+maintainer:          ktion23@gmail.com
+build-type:          Simple
+cabal-version:       >=1.10
 
- executable rekit
-   hs-source-dirs: src
-   main-is:           Main.hs
-   other-modules:
-              Asm
-              , Symbolic
-              , Util
-   build-depends:     base >=4.5 && <9.8
-                      , containers
-                      , bytestring
-                      , binary
-                      , hapstone >= 0.2.0.1
---   default-language:  Haskell2010
+executable rekit
+ hs-source-dirs: src
+ main-is:           Main.hs
+ other-modules:
+            Ast
+            , Asm
+            , Symbolic
+            , Util
+            , Lifter
+            , AstContext
+            , X86Sem
+
+ build-depends:     base >=4.5 && <9.8
+                    , containers
+                    , bytestring
+                    , binary
+                    , hapstone >= 0.2.0.1
+ default-language:  Haskell2010
+
+-- Test-Suite test-your-thing
+--    type:          detailed-0.9
+--    test-module:   Test.YourTestModule
+--    build-depends: base
+--                 , Cabal
+--                 , HUnit
